READ THE WHOLE COMMIT STATEMENT ON THIS BELOW
Highlight workings now, for voting which is nothign short of a miracle. A minor bug forces you to sometimes double click a vote to get it to go through.

CURRENTLY 50% and > 9 RULE IS ON. The lines that control this are 237-242! If you want to show immediate adding from game to view list you MUST swap the comments on the top and bottom lines of that range.

Signed-off-by: Thomas Van Dort <vanbarrington@gmail.com>